+++ clinicadl/callbacks/factory/__init__.py
@@ -1,0 +1,4 @@
+from .code_carbon import CodeCarbonCallback
+from .logger import LoggerCallback
+from .mlflow import MLFLOWCallback
+from .wandb import WandBCallback
