+++ clinicadl/callbacks/factory/base.py
@@ -1,0 +1,33 @@
+class Callback:
+    def __init__(self):
+        pass
+
+    def on_train_begin(self):
+        pass
+
+    def on_train_end(self):
+        pass
+
+    def on_epoch_begin(self):
+        pass
+
+    def on_epoch_end(self):
+        pass
+
+    def on_batch_begin(self):
+        pass
+
+    def on_batch_end(self):
+        pass
+
+    def on_loss_begin(self):
+        pass
+
+    def on_loss_end(self):
+        pass
+
+    def on_step_begin(self):
+        pass
+
+    def on_step_end(self):
+        pass
