+++ clinicadl/callbacks/factory/logger.py
@@ -1,0 +1,9 @@
+from .base import Callback
+
+
+class LoggerCallback(Callback):
+    def on_train_begin(self, **kwargs):
+        pass
+
+    def on_train_end(self, **kwargs):
+        pass
